+++ lib/merge/buildCommitMessage.js
@@ -3,7 +3,7 @@
 const debug = require('debug')('app:extractMessage')
 const getenv = require('getenv')
 
-const config = require('./config')
+const config = require('../config')
 
 const buildCommitMessage = function (title, description) {
   debug('Start')
