+++ lib/merge/index.js
@@ -5,7 +5,7 @@ const debug = require('debug')('app:merge')
 const checkForPermissions = require('./checkForPermissions')
 const checkForTrigger = require('./checkForTrigger')
 const checkForWip = require('./checkForWip')
-const buildCommitMessage = require('../buildCommitMessage')
+const buildCommitMessage = require('./buildCommitMessage')
 const buildResultMessage = require('./buildResultMessage')
 const isMergeComment = require('../isMergeComment')
